+++ src/xsarsea/windspeed/__init__.py
@@ -2,4 +2,5 @@ from .windspeed import *
 from . import gmfs
 from . import gmfs_impl
 from . import sarwing_luts
+from .utils import *
