+++ src/xsarsea/windspeed/gmfs.py
@@ -4,10 +4,105 @@ from ..xsarsea import logger, timing
 from functools import lru_cache
 from numba import njit, vectorize, guvectorize, float64, float32
 import xarray as xr
-from .sarwing_luts import lut_from_sarwing
-from .utils import registered_gmfs
 
 
+registered_gmfs = {}
+
+def register_gmf(name=None, inc_range=[17., 50.], wspd_range=[0.2, 50.], phi_range=None):
+    """TODO: docstring"""
+    def inner(func):
+        gmf_name = name or func.__name__
+
+        if not gmf_name.startswith('gmf_'):
+            raise ValueError("gmf function must start with 'gmf_'. Got %s" % gmf_name)
+
+        update = {
+            'gmf': func,
+            'inc_range': inc_range,
+            'wspd_range': wspd_range,
+            'phi_range': phi_range,
+        }
+        update = {k: update[k] for k in update.keys() if update[k] is not None}
+
+        if gmf_name not in registered_gmfs:
+            registered_gmfs[gmf_name] = {}
+
+        registered_gmfs[gmf_name].update(update)
+
+        return func
+
+    return inner
+
+@timing
+def _gmf_to_lut(name, inc_range=None, phi_range=None, wspd_range=None, allow_interp=True):
+
+    try:
+        cmod_attrs = registered_gmfs[name].copy()
+    except KeyError:
+        raise KeyError('gmf %s not available. (available: %s)' % (name, registered_gmfs.keys()))
+
+    inc_range = inc_range or cmod_attrs.pop('inc_range')
+    phi_range = phi_range or cmod_attrs.pop('phi_range', None)
+    wspd_range = wspd_range or cmod_attrs.pop('wspd_range')
+
+    inc_step_hr = 0.1
+    wspd_step_hr = 0.1
+    phi_step_hr = 1
+
+    inc_step_lr = 0.2
+    wspd_step_lr = 0.5
+    phi_step_lr = 1
+
+    if allow_interp:
+        inc_step = inc_step_lr
+        wspd_step = wspd_step_lr
+        phi_step = phi_step_lr
+    else:
+        inc_step = inc_step_hr
+        wspd_step = wspd_step_hr
+        phi_step = phi_step_hr
+
+    # 2*step, because we want to be sure to not have bounds conditions in interp
+    inc = np.arange(inc_range[0] - inc_step, inc_range[1] + 2*inc_step, inc_step)
+    wspd = np.arange(np.max([0, wspd_range[0] - wspd_step]), wspd_range[1] + 2*wspd_step, wspd_step)
+
+    try:
+        phi = np.arange(phi_range[0], phi_range[1] + 2*phi_step, phi_step)
+        dims = ['incidence', 'wspd', 'phi']
+        coords = {'incidence': inc, 'wspd': wspd, 'phi': phi}
+    except TypeError:
+        phi = None
+        dims = ['incidence', 'wspd']
+        coords = {'incidence': inc, 'wspd': wspd}
+
+
+    lut = xr.DataArray(
+        gmf(inc, wspd, phi, name),
+        dims=dims,
+        coords=coords
+    )
+
+    if allow_interp:
+        # interp to get high res
+        interp_kwargs = {}
+        interp_kwargs['incidence'] = np.arange(inc_range[0], inc_range[1] + inc_step_hr, inc_step_hr)
+        interp_kwargs['wspd'] = np.arange(wspd_range[0], wspd_range[1] + wspd_step_hr, wspd_step_hr)
+        if phi is not None:
+            interp_kwargs['phi'] = np.arange(phi_range[0], phi_range[1] + phi_step_hr, phi_step_hr)
+
+        lut = lut.interp(**interp_kwargs, kwargs=dict(bounds_error=True))
+
+    # crop lut to exact range
+    crop_cond = (lut.incidence >= inc_range[0]) & (lut.incidence <= inc_range[1])
+    crop_cond = crop_cond & (lut.wspd >= wspd_range[0]) & (lut.wspd <= wspd_range[1])
+    if phi is not None:
+        crop_cond = crop_cond & (lut.phi >= phi_range[0]) & (lut.phi <= phi_range[1])
+
+    lut = lut.where(crop_cond, drop=True)
+
+    # TODO add lut.attrs
+
+    return lut
 
 
 @timing
@@ -135,77 +230,6 @@ def gmf(inc, wspd, phi=None, name=None, numba=True):
 
     return sigma0_lin
 
-
-@timing
-def _gmf_to_lut(name, inc_range=None, phi_range=None, wspd_range=None, allow_interp=True):
-
-    try:
-        cmod_attrs = registered_gmfs[name].copy()
-    except KeyError:
-        raise KeyError('gmf %s not available. (available: %s)' % (name, registered_gmfs.keys()))
-
-    inc_range = inc_range or cmod_attrs.pop('inc_range')
-    phi_range = phi_range or cmod_attrs.pop('phi_range', None)
-    wspd_range = wspd_range or cmod_attrs.pop('wspd_range')
-
-    inc_step_hr = 0.1
-    wspd_step_hr = 0.1
-    phi_step_hr = 1
-
-    inc_step_lr = 0.2
-    wspd_step_lr = 0.5
-    phi_step_lr = 1
-
-    if allow_interp:
-        inc_step = inc_step_lr
-        wspd_step = wspd_step_lr
-        phi_step = phi_step_lr
-    else:
-        inc_step = inc_step_hr
-        wspd_step = wspd_step_hr
-        phi_step = phi_step_hr
-
-    # 2*step, because we want to be sure to not have bounds conditions in interp
-    inc = np.arange(inc_range[0] - inc_step, inc_range[1] + 2*inc_step, inc_step)
-    wspd = np.arange(np.max([0, wspd_range[0] - wspd_step]), wspd_range[1] + 2*wspd_step, wspd_step)
-
-    try:
-        phi = np.arange(phi_range[0], phi_range[1] + 2*phi_step, phi_step)
-        dims = ['incidence', 'wspd', 'phi']
-        coords = {'incidence': inc, 'wspd': wspd, 'phi': phi}
-    except TypeError:
-        phi = None
-        dims = ['incidence', 'wspd']
-        coords = {'incidence': inc, 'wspd': wspd}
-
-
-    lut = xr.DataArray(
-        gmf(inc, wspd, phi, name),
-        dims=dims,
-        coords=coords
-    )
-
-    if allow_interp:
-        # interp to get high res
-        interp_kwargs = {}
-        interp_kwargs['incidence'] = np.arange(inc_range[0], inc_range[1] + inc_step_hr, inc_step_hr)
-        interp_kwargs['wspd'] = np.arange(wspd_range[0], wspd_range[1] + wspd_step_hr, wspd_step_hr)
-        if phi is not None:
-            interp_kwargs['phi'] = np.arange(phi_range[0], phi_range[1] + phi_step_hr, phi_step_hr)
-
-        lut = lut.interp(**interp_kwargs, kwargs=dict(bounds_error=True))
-
-    # crop lut to exact range
-    crop_cond = (lut.incidence >= inc_range[0]) & (lut.incidence <= inc_range[1])
-    crop_cond = crop_cond & (lut.wspd >= wspd_range[0]) & (lut.wspd <= wspd_range[1])
-    if phi is not None:
-        crop_cond = crop_cond & (lut.phi >= phi_range[0]) & (lut.phi <= phi_range[1])
-
-    lut = lut.where(crop_cond, drop=True)
-
-    # TODO add lut.attrs
-
-    return lut
